requirements.txt
protobuf==4.25.1
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.29.0
